+++ sample/src/main/kotlin/com/github/panpf/sketch/sample/ui/notify/RemoteViewsTestFragment.kt
@@ -136,7 +136,6 @@ class RemoteViewsTestFragment : BaseToolbarBindingFragment<FragmentTestRemoteVie
                     RemoteViewsTarget(
                         remoteViews = remoteViews,
                         imageViewId = R.id.image,
-                        ignoreNullDrawable = true,
                         onUpdated = {
                             val notificationManager =
                                 context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager